code coverage setup for code climate

+++ lib/datashift/loaders/failure_data.rb
@@ -30,9 +30,11 @@ module DataShift
     end
 
     def destroy_failed_object
-      if load_object.respond_to?('destroy') && !load_object.new_record?
-        load_object.destroy
-      end if load_object
+      if load_object
+        if load_object.respond_to?('destroy') && !load_object.new_record?
+          load_object.destroy
+        end
+      end
     end
 
   end
